chore(login-api): [CHK-3720] rename package
rename client package to plural
